Move Toast API to its own namespace and set package to be an esm module
