+++ package.json
@@ -1,16 +1,21 @@
 {
   "name": "@fabric-ds/elements",
+  "type": "module",
   "version": "0.0.2",
   "description": "Custom elements for Fabric",
-  "exports": "./dist/index.js",
+  "exports": {
+    "./": "./dist/index.js",
+    "./toast": "./dist/api.js"
+  },
   "module": "dist/index.js",
   "files": [
     "dist/"
   ],
   "types": "dist/index.d.ts",
   "scripts": {
-    "prebuild": "rimraf packages/*/dist",
-    "build": "rimraf dist && tsc && NODE_ENV=production microbundle -i index.js -o dist/index.js --no-pkg-main --format modern",
+    "build:api": "npx esbuild ./packages/toast/api.js --outdir=dist --target=es2017 --bundle --sourcemap --format=esm",
+    "build:npm": "npx esbuild ./index.js --outdir=dist/ --target=es2017 --bundle --sourcemap --format=esm",
+    "build": "rimraf dist && tsc && npm run build:npm && npm run build:api",
     "format": "prettier --write . --ignore-path .gitignore",
     "lint": "npm run lint:format && npm run lint:eslint",
     "lint:format": "prettier --check . --ignore-path .gitignore",
